Eight items in the order they were planned, each with its own commit, PRE-vs-POST
differential over the whole scratchpad corpus, and a soundness pass.

  1    fork arm calling a suspendable task silently lost its body   [silent-wrong]
  7    negative unpacked-array low bound silently dropped a word    [silent-wrong]
  2    concurrent activations each keep their own frame-local dyn array
  3-6  generate + interface scopes: block-locals, string/queue decl-init
  8a   negative packed low bound (multi-packed inner was silent)    [silent-wrong]
  8b12 VCD declared range + $fmonitor/$fstrobe   (format 24 -> 25)

THREE RECORDED PREMISES TURNED OUT WRONG, and finding that out was most of the
work:

- the fork gap was filed as a deep scheduler rework. It was one comparison
  between two different bb numbering spaces. Identified by a discriminating
  experiment — padding the task CFG with nine dead `if`s moves the resume bb
  past the collision and the same design passes.
- "negative low bound = E4002, non-silent" held only when the negative index is
  touched explicitly. The pure-`foreach` form was silent (iverilog 3, vita 2, no
  diagnostic), and the multi-packed inner bound was silent outright — the clamp
  warning that made it look loud came from a sibling declaration.
- `allow_string_init=false` was documented as "those scopes do not run the
  string flush". They do. The flag was covering a scope leak: a string
  declaration's decl-time writes went into the module-scope pending list with a
  bare-name lvalue. Fixing the reason opened four queued items plus two
  follow-ons that were not on the list.

None of the six needed new machinery. Each was an existing rule applied where it
had been assumed: a numbering-space discriminator, a signed low bound through
the addressing helpers, the AUTOMATIC window's lifetime given to the dyn heap,
and pending lists keyed by the scope that declared them.

ONE REGRESSION, caught only by the 16k-design sweep: giving dyn arrays the
window's lifetime broke a fork arm reading its parent's array. A parked array is
ABSENT from the heap, not shared like a window handle, so the arm read X. Every
probe written for that slice passed; `snd/U4.sv` did not. Fixed and pinned.

FOUR PRE-EXISTING DEFECTS fell out along the way: `$bits` of a partially-indexed
array view (off by one), the engine's file-I/O base (same misread), `$readmemh`
with an explicit range on a non-zero-base memory, and a PANIC on a pathological
declared span — a crash is below loud on the ladder.

STILL LOUD, recorded rather than papered over: PART select of a negative-bound
net (same free function as the §2 part-select constant-folding gap — a candidate
to close together) · ports/formals with a negative bound (deliberate opt-in
asymmetry: width and select normalization must be enabled together) ·
hierarchical read of an interface queue element · the module-scope twin of the
block-local/net name collision that is now loud inside generate.

Also corrects a user-facing error doc: the F4004 table told users to "remove the
concurrency" for a fatal that is no longer a capability limit.

4613 tests green (+63) · clippy 0 · fmt 0 · format_version 25 (two append-only
sidecars; sim-ir unchanged, so the SimIr goldens are untouched) · every file
back under the 1000-line policy.

+++ crates/elaborate/src/array_formal.rs
@@ -411,10 +411,10 @@ impl Elaborator<'_> {
     pub(crate) fn array_formal_ext_dims(
         dims: &[(u32, u32, bool)],
         elem_w: u32,
-    ) -> Vec<(u32, u32, bool)> {
+    ) -> Vec<(i64, u32, bool)> {
         dims.iter()
-            .map(|&(lo, sz, _)| (lo, sz, false))
-            .chain(std::iter::once((0u32, elem_w, false)))
+            .map(|&(lo, sz, _)| (i64::from(lo), sz, false))
+            .chain(std::iter::once((0i64, elem_w, false)))
             .collect()
     }
